backend/requirements.txt
fastapi
uvicorn
yt-dlp
groq
python-multipart
python-dotenv
requests
